refactor(activerecord): move Cipher, Transaction to canonical locations
- Cipher: encryption/cipher.ts → encryption/cipher/aes256-gcm.ts
  (deleted old file, updated all imports)
- Transaction: extracted from transactions.ts → connection-adapters/abstract/transaction.ts
  (transactions.ts imports from new location)

Rollback remains re-exported from transactions.ts since dynamic imports
reference it there.

api:compare: 54/597 (9%), 8 misplaced (was 10).
