+++ packages/activerecord/src/encryption/cipher/aes256-gcm.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { Cipher } from "../cipher.js";
+import { Cipher } from "./aes256-gcm.js";
 import * as crypto from "crypto";
 
 function generateKey(): string {
