+++ packages/activerecord/src/transactions.ts
@@ -1,8 +1,8 @@
 import type { Base } from "./base.js";
-import type { DatabaseAdapter } from "./adapter.js";
 
 import { Rollback } from "./errors.js";
 export { Rollback };
+import { Transaction } from "./connection-adapters/abstract/transaction.js";
 
 // Track the currently-active transaction (if any) for after_commit/after_rollback callbacks
 let _currentTransaction: Transaction | null = null;
@@ -12,53 +12,6 @@ let _currentTransaction: Transaction | null = null;
  */
 export function currentTransaction(): Transaction | null {
   return _currentTransaction;
-}
-
-/**
- * Transaction — wraps adapter transactions with callbacks.
- *
- * Mirrors: ActiveRecord::Transactions
- */
-export class Transaction {
-  private adapter: DatabaseAdapter;
-  private _afterCommitCallbacks: Array<() => void | Promise<void>> = [];
-  private _afterRollbackCallbacks: Array<() => void | Promise<void>> = [];
-
-  constructor(adapter: DatabaseAdapter) {
-    this.adapter = adapter;
-  }
-
-  /**
-   * Register an after_commit callback.
-   */
-  afterCommit(fn: () => void | Promise<void>): void {
-    this._afterCommitCallbacks.push(fn);
-  }
-
-  /**
-   * Register an after_rollback callback.
-   */
-  afterRollback(fn: () => void | Promise<void>): void {
-    this._afterRollbackCallbacks.push(fn);
-  }
-
-  /**
-   * Execute the after_commit callbacks.
-   */
-  async runAfterCommitCallbacks(): Promise<void> {
-    for (const fn of this._afterCommitCallbacks) {
-      await fn();
-    }
-  }
-
-  /**
-   * Execute the after_rollback callbacks.
-   */
-  async runAfterRollbackCallbacks(): Promise<void> {
-    for (const fn of this._afterRollbackCallbacks) {
-      await fn();
-    }
-  }
 }
 
 /**
